Rename notif to notif_gomail and get the mailer from config
