+++ routes/users.js
@@ -13,6 +13,9 @@ const logger = winston.loggers.get('gomngr');
 var CONFIG = require('config');
 var GENERAL_CONFIG = CONFIG.general;
 
+const MAILER = CONFIG.general.mailer;
+const MAIL_CONFIG = CONFIG[MAILER];
+
 var plugins = CONFIG.plugins;
 if(plugins === undefined){
     plugins = [];
@@ -23,7 +26,7 @@ var plugins_info = [];
 var cookieParser = require('cookie-parser');
 
 var goldap = require('../routes/goldap.js');
-var notif = require('../routes/notif.js');
+var notif = require('../routes/notif_'+MAILER+'.js');
 var utils = require('../routes/utils.js');
 
 // var get_ip = require('ipware')().get_ip;
@@ -42,8 +45,6 @@ var STATUS_PENDING_EMAIL = 'Waiting for email approval';
 var STATUS_PENDING_APPROVAL = 'Waiting for admin approval';
 var STATUS_ACTIVE = 'Active';
 var STATUS_EXPIRED = 'Expired';
-
-var MAIL_CONFIG = CONFIG.gomail;
 
 const runningEnv = process.env.NODE_ENV || 'prod';
